feat: move fix after JSON.stringify
