+++ front/temporal/agent_loop/lib/tool_arguments.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 
-import { fixCorruptedUnicode, parseToolArguments } from "./tool_arguments";
+import { fixCorruptedUnicode } from "./tool_arguments";
 
 describe("fixCorruptedUnicode", () => {
   describe("corrupted Latin-1 Supplement characters", () => {
@@ -234,155 +234,6 @@ describe("fixCorruptedUnicode", () => {
     it("should preserve newlines, tabs, and other whitespace", () => {
       expect(fixCorruptedUnicode("Line1\x00e9\nLine2\x00e0\tTab")).toBe(
         "Line1é\nLine2à\tTab"
-      );
-    });
-  });
-});
-
-describe("parseToolArguments", () => {
-  describe("with corrupted Unicode", () => {
-    it("should parse JSON with corrupted Unicode characters", () => {
-      const input =
-        '{"query":"Acme Corp IA g\x00e9n\x00e9rative relances clients factures IA"}';
-      const result = parseToolArguments(input, "test_tool");
-      expect(result).toEqual({
-        query: "Acme Corp IA générative relances clients factures IA",
-      });
-    });
-
-    it("should parse complex JSON with multiple corrupted fields", () => {
-      const input =
-        '{"query":"TaxConsult France IA g\x00e9n\x00e9rative fiscalit\x00e9","relativeTimeFrame":"all"}';
-      const result = parseToolArguments(input, "test_tool");
-      expect(result).toEqual({
-        query: "TaxConsult France IA générative fiscalité",
-        relativeTimeFrame: "all",
-      });
-    });
-
-    it("should handle nested objects with corruption", () => {
-      const input = '{"data":{"text":"tr\x00e8s bien","count":42}}';
-      const result = parseToolArguments(input, "test_tool");
-      expect(result).toEqual({
-        data: {
-          text: "très bien",
-          count: 42,
-        },
-      });
-    });
-
-    it("should handle arrays with corrupted strings", () => {
-      const input = '{"items":["caf\x00e9","th\x00e9","cr\x00e8me"]}';
-      const result = parseToolArguments(input, "test_tool");
-      expect(result).toEqual({
-        items: ["café", "thé", "crème"],
-      });
-    });
-  });
-
-  describe("without corruption", () => {
-    it("should parse normal JSON correctly", () => {
-      const input = '{"query":"Hello World","count":5}';
-      const result = parseToolArguments(input, "test_tool");
-      expect(result).toEqual({
-        query: "Hello World",
-        count: 5,
-      });
-    });
-
-    it("should parse already-correct Unicode", () => {
-      const input = '{"query":"café et thé"}';
-      const result = parseToolArguments(input, "test_tool");
-      expect(result).toEqual({
-        query: "café et thé",
-      });
-    });
-
-    it("should return empty object for empty string", () => {
-      const result = parseToolArguments("", "test_tool");
-      expect(result).toEqual({});
-    });
-
-    it("should return empty object for whitespace-only string", () => {
-      const result = parseToolArguments("  \n  ", "test_tool");
-      expect(result).toEqual({});
-    });
-  });
-
-  describe("error cases", () => {
-    it("should throw on invalid JSON", () => {
-      expect(() => parseToolArguments('{"invalid": }', "test_tool")).toThrow(
-        /Failed to parse arguments in call to tool 'test_tool'/
-      );
-    });
-
-    it("should throw on non-object JSON", () => {
-      expect(() => parseToolArguments('"just a string"', "test_tool")).toThrow(
-        /Tool call arguments must be an object/
-      );
-
-      expect(() => parseToolArguments("123", "test_tool")).toThrow(
-        /Tool call arguments must be an object/
-      );
-
-      expect(() => parseToolArguments('["array"]', "test_tool")).toThrow(
-        /Tool call arguments must be an object/
-      );
-    });
-
-    it("should throw on null", () => {
-      expect(() => parseToolArguments("null", "test_tool")).toThrow(
-        /Tool call arguments must be an object/
-      );
-    });
-  });
-
-  describe("integration with real GPT-5 examples", () => {
-    it("should handle Example 1", () => {
-      const input =
-        '{"query":"Acme Corp IA g\x00e9n\x00e9rative relances clients factures IA"}';
-      const result = parseToolArguments(input, "search_tool");
-      expect(result.query).toBe(
-        "Acme Corp IA générative relances clients factures IA"
-      );
-    });
-
-    it("should handle Example 2", () => {
-      const input =
-        '{"query":"AutoInsure Co - ton \x00e9ditorial et voix de marque pour contenus daitoriel assurance et services, ' +
-        "style p\x00e9dagogique, accessible, rassurant, orient\x00e9 grand public. Points \x00e0 respecter : " +
-        "clart\x00e9, phrases courtes, vocabulaire simple mais pr\x00e9cis, posture experte et bienveillante, " +
-        "ancrage assurance auto / habitation, neutralit\x00e9 commerciale, conseils pratiques. " +
-        "Mots \x00e0 privil\x00e9gier / \x00e9viter ? Longueur attendue Q&A 80-150 mots. " +
-        'Angle H3 : \\"Devis gratuit ou payant : que dit la loi ?\\" dans H2 \\"Questions fr\x00e9quentes\\".",' +
-        '"relativeTimeFrame":"all"}';
-      const result = parseToolArguments(input, "search_tool");
-      expect(result.query).toContain("pédagogique");
-      expect(result.query).toContain("orienté");
-      expect(result.query).toContain("à respecter");
-      expect(result.query).toContain("clarté");
-      expect(result.query).toContain("précis");
-      expect(result.query).toContain("neutralité");
-      expect(result.query).toContain("privilégier");
-      expect(result.query).toContain("éviter");
-      expect(result.query).toContain("fréquentes");
-    });
-
-    it("should handle Example 3", () => {
-      const input =
-        '{"query":"TaxConsult France IA g\x00e9n\x00e9rative fiscalit\x00e9 copilot"}';
-      const result = parseToolArguments(input, "search_tool");
-      expect(result.query).toBe(
-        "TaxConsult France IA générative fiscalité copilot"
-      );
-    });
-
-    it("should handle Example 4", () => {
-      const input =
-        '{"query":"AccountPro IA g\x00e9n\x00e9rative experts\x00d7\x00b7comptables France"}';
-      const result = parseToolArguments(input, "search_tool");
-      expect(result.query).toBe(
-        "AccountPro IA générative experts×·comptables France"
       );
     });
   });
